rename ut_cnstitsel.py to constituent_selection.py
